chore: Added tenant-aware integ tests for IdP config management
