+++ FirebaseAdmin/FirebaseAdmin.IntegrationTests/Auth/SamlProviderConfigFixture.cs
@@ -1,0 +1,50 @@
+using System;
+using System.Collections.Generic;
+using FirebaseAdmin.Auth;
+using FirebaseAdmin.Auth.Providers;
+
+namespace FirebaseAdmin.IntegrationTests.Auth
+{
+    public abstract class SamlProviderConfigFixture<T> : IDisposable
+    where T : AbstractFirebaseAuth
+    {
+        internal static readonly IList<string> X509Certificates = new List<string>()
+        {
+            "-----BEGIN CERTIFICATE-----\nMIICZjCCAc+gAwIBAgIBADANBgkqhkiG9w0BAQ0FADBQMQswCQYDVQQGEwJ1czEL\nMAkGA1UECAwCQ0ExDTALBgNVBAoMBEFjbWUxETAPBgNVBAMMCGFjbWUuY29tMRIw\nEAYDVQQHDAlTdW5ueXZhbGUwHhcNMTgxMjA2MDc1MTUxWhcNMjgxMjAzMDc1MTUx\nWjBQMQswCQYDVQQGEwJ1czELMAkGA1UECAwCQ0ExDTALBgNVBAoMBEFjbWUxETAP\nBgNVBAMMCGFjbWUuY29tMRIwEAYDVQQHDAlTdW5ueXZhbGUwgZ8wDQYJKoZIhvcN\nAQEBBQADgY0AMIGJAoGBAKphmggjiVgqMLXyzvI7cKphscIIQ+wcv7Dld6MD4aKv\n7Jqr8ltujMxBUeY4LFEKw8Terb01snYpDotfilaG6NxpF/GfVVmMalzwWp0mT8+H\nyzyPj89mRcozu17RwuooR6n1ofXjGcBE86lqC21UhA3WVgjPOLqB42rlE9gPnZLB\nAgMBAAGjUDBOMB0GA1UdDgQWBBS0iM7WnbCNOnieOP1HIA+Oz/ML+zAfBgNVHSME\nGDAWgBS0iM7WnbCNOnieOP1HIA+Oz/ML+zAMBgNVHRMEBTADAQH/MA0GCSqGSIb3\nDQEBDQUAA4GBAF3jBgS+wP+K/jTupEQur6iaqS4UvXd//d4vo1MV06oTLQMTz+rP\nOSMDNwxzfaOn6vgYLKP/Dcy9dSTnSzgxLAxfKvDQZA0vE3udsw0Bd245MmX4+GOp\nlbrN99XP1u+lFxCSdMUzvQ/jW4ysw/Nq4JdJ0gPAyPvL6Qi/3mQdIQwx\n-----END CERTIFICATE-----\n",
+            "-----BEGIN CERTIFICATE-----\nMIICZjCCAc+gAwIBAgIBADANBgkqhkiG9w0BAQ0FADBQMQswCQYDVQQGEwJ1czEL\nMAkGA1UECAwCQ0ExDTALBgNVBAoMBEFjbWUxETAPBgNVBAMMCGFjbWUuY29tMRIw\nEAYDVQQHDAlTdW5ueXZhbGUwHhcNMTgxMjA2MDc1ODE4WhcNMjgxMjAzMDc1ODE4\nWjBQMQswCQYDVQQGEwJ1czELMAkGA1UECAwCQ0ExDTALBgNVBAoMBEFjbWUxETAP\nBgNVBAMMCGFjbWUuY29tMRIwEAYDVQQHDAlTdW5ueXZhbGUwgZ8wDQYJKoZIhvcN\nAQEBBQADgY0AMIGJAoGBAKuzYKfDZGA6DJgQru3wNUqv+S0hMZfP/jbp8ou/8UKu\nrNeX7cfCgt3yxoGCJYKmF6t5mvo76JY0MWwA53BxeP/oyXmJ93uHG5mFRAsVAUKs\ncVVb0Xi6ujxZGVdDWFV696L0BNOoHTfXmac6IBoZQzNNK4n1AATqwo+z7a0pfRrJ\nAgMBAAGjUDBOMB0GA1UdDgQWBBSKmi/ZKMuLN0ES7/jPa7q7jAjPiDAfBgNVHSME\nGDAWgBSKmi/ZKMuLN0ES7/jPa7q7jAjPiDAMBgNVHRMEBTADAQH/MA0GCSqGSIb3\nDQEBDQUAA4GBAAg2a2kSn05NiUOuWOHwPUjW3wQRsGxPXtbhWMhmNdCfKKteM2+/\nLd/jz5F3qkOgGQ3UDgr3SHEoWhnLaJMF4a2tm6vL2rEIfPEK81KhTTRxSsAgMVbU\nJXBz1md6Ur0HlgQC7d1CHC8/xi2DDwHopLyxhogaZUxy9IaRxUEa2vJW\n-----END CERTIFICATE-----\n",
+        };
+
+        public SamlProviderConfigFixture()
+        {
+            IntegrationTestUtils.EnsureDefaultApp();
+
+            var providerId = $"saml.{AuthIntegrationUtils.GetRandomIdentifier()}";
+            var args = new SamlProviderConfigArgs
+            {
+                ProviderId = providerId,
+                DisplayName = "SAML_DISPLAY_NAME",
+                Enabled = true,
+                IdpEntityId = "IDP_ENTITY_ID",
+                SsoUrl = "https://example.com/login",
+                X509Certificates = new List<string> { X509Certificates[0] },
+                RpEntityId = "RP_ENTITY_ID",
+                CallbackUrl = "https://projectId.firebaseapp.com/__/auth/handler",
+            };
+            this.ProviderConfig = this.Auth.CreateProviderConfigAsync(args).Result;
+        }
+
+        public abstract T Auth { get; }
+
+        public SamlProviderConfig ProviderConfig { get; set; }
+
+        public string ProviderId => this.ProviderConfig.ProviderId;
+
+        public virtual void Dispose()
+        {
+            if (this.ProviderConfig != null)
+            {
+                this.Auth.DeleteProviderConfigAsync(this.ProviderConfig.ProviderId).Wait();
+            }
+        }
+    }
+}
